+++ lib/massive_sitemap.rb
@@ -6,6 +6,7 @@ require "massive_sitemap/version"
 require 'massive_sitemap/writer/file'
 require 'massive_sitemap/writer/gzip_file'
 require 'massive_sitemap/builder/rotating'
+require 'massive_sitemap/builder/index'
 
 # Page at -> <base_url>
 # http://example.de/dir/
@@ -16,20 +17,19 @@ require 'massive_sitemap/builder/rotating'
 # Save at -> <document_full>
 # /root/dir/ ->  <document_root>/<document_path>
 
-# require 'massive_sitemap/builder/index'
-
 module MassiveSitemap
   DEFAULTS = {
     # builder
-    :base_url        => nil,
-    :indent_by       => 2,
+    :base_url               => nil,
+    :indent_by              => 2,
 
     # writer
-    :document_full   => '.',
-    :force_overwrite  => false,
-
+    :document_full          => '.',
+    :force_overwrite        => false,
+    :sitemap_filename       => "sitemap.xml",
+    :index_filename         => "sitemap_index.xml",
     # writer gzip
-    :gzip            => false,
+    :gzip                   => false,
   }
 
   def generate(options = {}, &block)
@@ -43,12 +43,34 @@ module MassiveSitemap
 
     @writer_class = @options[:gzip] ? Writer::GzipFile : Writer::File
 
-    @writer = @writer_class.new "sitemap.xml", @options
+    generate_sitemap(&block)
+  end
+  module_function :generate
+
+  def generate_sitemap(&block)
+    @writer = @writer_class.new @options[:sitemap_filename], @options
     @builder = Builder::Rotating.new(@writer, @options)
     instance_eval(&block) if block
     @builder.close!
+    self
   end
-  module_function :generate
+  module_function :generate_sitemap
+
+  # Create a sitemap index document
+  def generate_index(files = nil)
+    ext     = @options[:gzip] ? "xml.gz" : "xml"
+    files ||= Dir[File.join(@options[:document_full], "*.#{ext}")]
+
+    @writer = @writer_class.new @options[:index_filename], @options.merge(:force_overwrite => true)
+    Builder::Index.new(@writer, @options.merge(:base_url => "http://test.de")) do
+      files.each do |path|
+        next if path.include?(@options[:index_filename])
+        add ::File.basename(path), :last_modified => File.stat(path).mtime
+      end
+    end
+    self
+  end
+  module_function :generate_index
 
   def add(path, attrs = {})
     @builder.add(path, attrs)